Refactoring common code into PythonCommon
